moved auth to middleware, cleaned code up. fixed repeated urls on db

+++ internal/database/feedfollows.sql.go
@@ -49,10 +49,16 @@ func (q *Queries) CreateFeedFollow(ctx context.Context, arg CreateFeedFollowPara
 const deleteFeedFollow = `-- name: DeleteFeedFollow :exec
 DELETE FROM feedfollow
 WHERE id = $1
+    AND user_id = $2
 `
 
-func (q *Queries) DeleteFeedFollow(ctx context.Context, id uuid.UUID) error {
-	_, err := q.db.ExecContext(ctx, deleteFeedFollow, id)
+type DeleteFeedFollowParams struct {
+	ID     uuid.UUID
+	UserID uuid.UUID
+}
+
+func (q *Queries) DeleteFeedFollow(ctx context.Context, arg DeleteFeedFollowParams) error {
+	_, err := q.db.ExecContext(ctx, deleteFeedFollow, arg.ID, arg.UserID)
 	return err
 }
